% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comsir.R
\name{comsir}
\alias{comsir}
\title{COM-SIR method}
\arguments{
\item{x}{Intrinsic rate of increase in effort}

\item{ct}{A time series of catch}

\item{yrs}{A time series of years associated with the catch}

\item{k}{Stock size at carrying capacity}

\item{r}{Intrinsic rate of population growth}

\item{a}{Fraction of K at bioeconomic equilibrium TODO: is this correct?}

\item{start_r}{A numeric vector of length 2 giving the lower and upper
bounds on the population growth rate parameter. This can either be
specified manually or by translating resiliency categories via the function
\code{\link{resilience}}}

\item{mink}{Minimum possible stock size at carrying capacity}

\item{maxk}{Max possible stock size at carrying capacity}

\item{logk}{Logical:}

\item{norm_k}{Logical:}

\item{norm_r}{Logical:}

\item{norm_a}{Logical:}

\item{norm_x}{Logical:}

\item{nsim}{Number of iterations to run before sampling}

\item{cv}{Coefficient of variation on ... TODO}

\item{logistic_model}{Logical:}

\item{obs}{Logical:}

\item{n_posterior}{Number of posterior samples to draw}
}
\description{
Catch-only model with sample importance resampling.
}
\examples{
# TODO K and r values?
x <- comsir(ct = blue_gren$ct, yrs = blue_gren$yrs, k = 800, r = 0.6)
head(x)
par(mfrow = c(1, 2))
hist(x$BoverBmsy)
with(x$posterior, plot(r, k))
}

