% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prm.R
\name{fit_prm}
\alias{fit_prm}
\title{Fit panel regression}
\usage{
fit_prm(dat)
}
\arguments{
\item{dat}{A data frame created by \code{\link{format_prm}}}
}
\value{
A linear model
}
\description{
Fit panel regression
}
\examples{
m <- fit_prm(ram_prm_dat)
}
\references{
Costello, C., D. Ovando, R. Hilborn, S. D. Gaines, O. Deschenes,
and S. E. Lester. 2012. Status and Solutions for the World's Unassessed
Fisheries. Science 338:517-520.
}
\seealso{
\code{\link{format_prm}}, \code{\link{predict_prm}}

Other prm: \code{\link{format_prm}};
  \code{\link{predict_prm}}
}
