% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prm.R
\name{format_prm}
\alias{format_prm}
\title{Format a series of predictors for a panel regression}
\usage{
format_prm(year, catch, bbmsy, species_cat)
}
\arguments{
\item{year}{A numeric vector of years}

\item{catch}{A numeric vector of catches}

\item{bbmsy}{A numeric vector of B/B_MSY}

\item{species_cat}{A single character value of a
species category. Can be from any set of species categories as long as the
same set are used in model fitting and extrapolation.}
}
\value{
A data frame formatted for use with \code{\link{fit_prm}} or
  \code{\link{predict_prm}}
}
\description{
Format a series of predictors for a panel regression
}
\examples{
d <- dplyr::inner_join(ramts, spp_categories, by = "scientificname")
ram_prm_dat <- plyr::ddply(d, "stockid", function(x) {
  format_prm(year = x$year, catch = x$catch, bbmsy = x$bbmsy_ram,
    species_cat = x$spp_category[1L])
})
}
\references{
Costello, C., D. Ovando, R. Hilborn, S. D. Gaines, O. Deschenes,
and S. E. Lester. 2012. Status and Solutions for the World's Unassessed
Fisheries. Science 338:517-520.
}
\seealso{
Other prm: \code{\link{fit_prm}}; \code{\link{predict_prm}}
}
