% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prm.R
\name{predict_prm}
\alias{predict_prm}
\title{Predict from a panel regression model}
\usage{
predict_prm(newdata, model = datalimited::ram_prm_model, ci = FALSE,
  level = 0.95)
}
\arguments{
\item{newdata}{A data frame to predict on that has been formatted with
\code{\link{format_prm}}.}

\item{model}{A linear model to predict from built from \code{\link{fit_prm}}.
Defaults to a cached version of a model fit to the RAM database,
\code{\link{ram_prm_model}}.}

\item{ci}{Should confidence intervals on B/B_MSY be returned?}

\item{level}{Confidence interval level}
}
\value{
If \code{ci = FALSE}, a vector of predictions of B/B_MSY. If \code{ci
  = TRUE}, a data frame with columns for B/B_MSY \code{fit} and lower and
  upper confidence intervals \code{lower}, \code{upper}. Note that the model
  is fitted to log(B/B_MSY) and the output from \code{predict_prm} is
  exponentiated so the prediction represents an estimate of median B/B_MSY.
}
\description{
Predict from a panel regression model
}
\examples{
d <- subset(ram_prm_dat, stockid == "BGRDRSE")
x <- predict_prm(d)
plot(x)

library("ggplot2")
x <- predict_prm(d, ci = TRUE)
x$year <- rev(d$years_back)
ggplot(x, aes(year, fit)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = "#00000040") +
  geom_line() +
  ylab(expression(B/B[MSY]))
}
\references{
Costello, C., D. Ovando, R. Hilborn, S. D. Gaines, O. Deschenes,
and S. E. Lester. 2012. Status and Solutions for the World's Unassessed
Fisheries. Science 338:517-520.
}
\seealso{
Other prm: \code{\link{fit_prm}}; \code{\link{format_prm}}
}
