% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm.R
\name{prm}
\alias{fit_prm}
\alias{format_prm}
\alias{predict_prm}
\alias{prm}
\title{Modified panel regression model}
\usage{
format_prm(year, catch, bbmsy, species_cat)

fit_prm(dat, eqn = log(bbmsy) ~ max_catch + mean_scaled_catch + scaled_catch +
  scaled_catch1 + scaled_catch2 + scaled_catch3 + scaled_catch4 + species_cat +
  catch_to_rolling_max + time_to_max + years_back + initial_slope - 1,
  type = c("lm", "gbm"), ...)

predict_prm(newdata, model = datalimited::ram_prm_model, ci = FALSE)
}
\arguments{
\item{year}{A numeric vector of years}

\item{catch}{A numeric vector of catches}

\item{bbmsy}{A numeric vector of B/B_MSY}

\item{species_cat}{A single character value of a
species category. Can be from any set of species categories as long as the
same set are used in model fitting and extrapolation.}

\item{dat}{A data frame created by \code{\link{format_prm}}}

\item{eqn}{A formula describing the regression}

\item{type}{Either linear model (\code{"lm"}; the original method) or a
\code{\link[gbm]{gbm}} model (\code{"gbm"})}

\item{...}{Anything extra to pass to \code{\link[gbm]{gbm}}}

\item{newdata}{A data frame to predict on that has been formatted with
\code{\link{format_prm}}.}

\item{model}{A linear model to predict from built from \code{\link{fit_prm}}.
Defaults to a cached version of a model fit to the RAM database,
\code{\link{ram_prm_model}}.}

\item{ci}{Should confidence intervals on B/B_MSY be returned?}
}
\value{
\code{format_prm}: A data frame formatted for use with \code{fit_prm} or
  \code{predict_prm}

\code{fit_prm}: A linear model for use with \code{predict_prm}

\code{predict_prm}: If \code{ci = FALSE}, a vector of predictions of
  B/B_MSY. If \code{ci = TRUE}, a data frame.
  Note that the model is fitted to log(B/B_MSY) and the output
  from \code{predict_prm} is exponentiated so the prediction represents an
  estimate of median B/B_MSY.
}
\description{
Format for, fit, and predict from a panel regression model. The
\pkg{datalimited} package includes cached data that can be used with
\code{fit_prm} and a cached model that can be used with \code{predict_prm}.
These functions can also be used with new data.

\code{format_prm}: Format a series of predictors for a panel
  regression

\code{fit_prm}: Fit a panel regression

\code{predict_prm}: Predict from a panel regression model
}
\examples{
# combine two built in datasets:
# d <- dplyr::inner_join(ram_ts, spp_categories, by = "scientificname")

# ram_prm_dat is built into the package; it can be created with:
# ram_prm_dat <- plyr::ddply(d, "stockid", function(x) {
#   format_prm(year = x$year, catch = x$catch, bbmsy = x$bbmsy_ram,
#     species_cat = x$spp_category[1L])
# })

# ram_prm_model is built into the package, it is created in the same manner
# as this:
# m <- fit_prm(ram_prm_dat)

# now predict B/Bmsy:
d <- subset(ram_prm_dat, stockid == "BGRDRSE")
x <- predict_prm(d) # use the built in model
plot(x)

# with confidence intervals:
library("ggplot2")
x <- predict_prm(d, ci = TRUE)
ggplot(x, aes(year, bbmsy_q50)) +
  geom_ribbon(aes(ymin = bbmsy_q2.5, ymax = bbmsy_q97.5), fill = "#00000040") +
  geom_line() +
  ylab(expression(B/B[MSY]))
}
\references{
Costello, C., D. Ovando, R. Hilborn, S. D. Gaines, O. Deschenes,
and S. E. Lester. 2012. Status and Solutions for the World's Unassessed
Fisheries. Science 338:517-520.
}

