% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{schaefer_cmsy}
\alias{schaefer_cmsy}
\title{Schaefer production model for catch-MSY method}
\usage{
schaefer_cmsy(r_lim, k_lim, sig_r, startbt, yr, ct, interyr_index,
  prior_log_mean, prior_log_sd, interbio_lim, reps)
}
\arguments{
\item{r_lim}{Numeric vector of lower and upper intrinsic population growth
rates}

\item{k_lim}{Numeric vector of lower and upper carrying capacities}

\item{sig_r}{Standard deviation of process noise}

\item{startbt}{A vector of possible starting biomasses to loop over}

\item{yr}{Numeric vector of years}

\item{ct}{Numeric vector of catch}

\item{interyr_index}{Index of the interim year within time series for which
biomass estimate is available}

\item{prior_log_mean}{Prior log mean}

\item{prior_log_sd}{Prior log sd}

\item{interbio_lim}{A numeric vector that gives the lower and upper biomass
limits in the interim year}

\item{reps}{Number of repititions}
}
\description{
Schaefer production model for catch-MSY method
}
\examples{
schaefer_cmsy(c(0.1, 0.3), c(80, 100), sig_r = 0.1,
  startbt = seq(0.2, 0.6, by = 0.05), yr = 1:10, ct = rnorm(10),
  prior_log_mean = 0.3, prior_log_sd = 0.1, interyr_index = 2,
  interbio_lim = c(0, 1), reps = 10L)
}
\references{
Martell, S., & Froese, R. (2013). A simple method for estimating MSY from
catch and resilience. Fish and Fisheries, 14(4), 504-514.
\url{http://doi.org/10.1111/j.1467-2979.2012.00485.x}
}
\seealso{
\code{\link{cmsy}}
}

