% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ram_data.R
\docType{data}
\name{spp_categories}
\alias{spp_categories}
\title{Species names and categories for use with the RAM database}
\description{
Cached from an analysis post Rosenberg et al. (2014).
}
\references{
Ricard, D., Minto, C., Jensen, O. P., & Baum, J. K. (2012). Examining the
knowledge base and status of commercially exploited marine species with the
RAM Legacy Stock Assessment Database. Fish and Fisheries, 13(4), 3 80-398.
\url{http://doi.org/10.1111/j.1467-2979.2011.00435.x}.
\url{http://ramlegacy.org}

Rosenberg, A. A., M. J. Fogarty, A. B. Cooper, M. Dickey-Collas, E. A.
Fulton, N. L. Gutierrez, K. J. W. Hyde, K. M. Kleisner, C. Longo, C. V.
Minte-Vera, C. Minto, I. Mosqueira, G. C. Osio, D. Ovando, E. R. Selig, J. T.
Thorson, and Y. Ye. 2014. Developing new approaches to global stock status
assessment and fishery production potential of the seas. FAO Fisheries and
Aquaculture Circular, Rome, Italy.
}
\keyword{data}

