% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscom.R
\name{sscom}
\alias{sscom}
\title{SSCOM model}
\usage{
sscom(yr, ct, start_r = resilience("unknown"), NburninPrelim = 1000,
  NiterPrelim = 2000, NthinPrelim = 1, NchainsPrelim = 100,
  NburninJags = 1e+06, NiterJags = 2e+06, NthinJags = 1000, Nchains = 3,
  return_jags = FALSE)
}
\arguments{
\item{yr}{Years}

\item{ct}{Catch time series}

\item{start_r}{Lower and upper intrinsic population growth rates}

\item{NburninPrelim}{Number of warmup iterations during the preliminary phase}

\item{NiterPrelim}{Number of preliminary iterations}

\item{NthinPrelim}{Thinning of preliminary iterations}

\item{NchainsPrelim}{Number of chains for preliminary iterations}

\item{NburninJags}{Number of warmup iterations for main sampling phase}

\item{NiterJags}{Number of sampling iterations for main sampling phase}

\item{NthinJags}{Thining for main sampling phase}

\item{Nchains}{Number of chains for the main sampling phase}

\item{return_jags}{Logical: should the JAGS output object be returned at the
end as well. If \code{TRUE} then a list with elements \code{jags} and
\code{output} will be returned with the JAGS and SSCOM output, respectively.}
}
\description{
State-space catch-only model from Thorson et al. (2013).
}
\examples{
\dontrun{
x <- sscom(blue_gren$yr, blue_gren$ct,
 NburninPrelim = 1e3,
 NiterPrelim = 2e3,
 NthinPrelim = 1,
 NchainsPrelim = 20,
 NburninJags = 1e3,
 NiterJags = 3e3,
 NthinJags = 2,
 Nchains = 3, return_jags = TRUE)
 }
}
\references{
Thorson, J. T., C. Minto, C. V. Minte-Vera, K. M. Kleisner and C. Longo,
2013. A new role for effort dynamics in the theory of harvested populations
and data-poor stock assessment. Can. J. Fish. Aquat. Sci. 70(12):1829-1844.
}

