% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarize_bbmsy.R
\name{summarize_bbmsy}
\alias{summarize_bbmsy}
\title{Summarize B/B_MSY values}
\usage{
summarize_bbmsy(bbmsy, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), log = TRUE,
  ...)
}
\arguments{
\item{bbmsy}{A numeric matrix of B/B_MSY values. Columns should represent
samples of B/B_MSY and rows should represent years.}

\item{probs}{A numeric vector of quantile probabilities.}

\item{log}{Logical: should the mean and standard deviation be calculated on
the log scale and then exponentiated at the end?}

\item{...}{Other parameters to pass to \code{quantile}, \code{sd}, and
\code{mean}. For example, \code{na.rm = TRUE}.}
}
\value{
A data frame: rows are years and columns represent quantile, mean,
  and standard deviation values of B/B_MSY.
}
\description{
Summarize B/B_MSY values
}

