% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cmsy.R
\name{cmsy}
\alias{cmsy}
\title{Catch-msy method}
\usage{
cmsy(yr, ct, prior_log_mean, prior_log_sd, interyr_index = 2,
  interbio = c(0, 1), bio_step = 0.05, resilience = c(NA, "Very low",
  "Low", "Medium", "High"), start_k = c(max(ct), 50 * max(ct)),
  startbio = if (ct[1]/max(ct) < 0.2) c(0.5, 0.9) else c(0.2, 0.6),
  sig_r = 0.05, reps = 10000)
}
\arguments{
\item{yr}{}

\item{ct}{}

\item{prior_log_mean}{}

\item{prior_log_sd}{}

\item{interyr_index}{}

\item{interbio}{}

\item{bio_step}{}

\item{resilience}{}

\item{start_k}{}

\item{startbio}{}

\item{sig_r}{}

\item{reps}{}

\item{finalbio}{}
}
\description{
Catch-msy method
}
\examples{
x <- cmsy(1:30, ct = rlnorm(30), prior_log_mean = 1, prior_log_sd = 0.5)
head(x)
}
\seealso{
\code{\link{shaefer_cmsy}}
}

