% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cmsy.R
\name{cmsy}
\alias{cmsy}
\title{Catch-MSY method}
\usage{
cmsy(yr, ct, interyr_index = 2L, interbio = c(0, 1), bio_step = 0.05,
  start_r = resilience(NA), start_k = c(max(ct), 50 * max(ct)),
  startbio = if (ct[1]/max(ct) < 0.2) c(0.5, 0.9) else c(0.2, 0.6),
  sig_r = 0.05, reps = 1e+05, revise_bounds = TRUE, finalbio = if
  (ct[length(ct)]/max(ct) > 0.5) c(0.3, 0.7) else c(0.01, 0.4))
}
\arguments{
\item{yr}{Numeric vector of years}

\item{ct}{Numeric vector of catch}

\item{interyr_index}{Index of the interim year within time series for which
biomass estimate is available}

\item{interbio}{A numeric vector that gives the lower and upper biomass
limits in the interim year}

\item{bio_step}{Step size between lower and upper biomass limits in the iterim year}

\item{start_r}{A numeric vector of length 2 giving the lower and upper
bounds on the population growth rate parameter. This can either be
specified manually or by translating resiliency categories via the function
\code{\link{resilience}}}

\item{start_k}{Numeric vector of length 2 giving the lower and upper starting
bounds on stock biomass at carrying capacity}

\item{startbio}{Starting biomass depletion}

\item{sig_r}{Standard deviation of process noise}

\item{reps}{Number of repititions to run the sampling}

\item{revise_bounds}{Should the bounds on r and k be revised after fitting
the algorithm once? The algorithm will then fit a second time with the
revised bounds.}

\item{finalbio}{Maximum and minimum biomass depletion in the final year}
}
\value{
A list containing a matrix \code{biomass}; \code{bmsy}, \code{msy},
and \code{mean_ln_msy} management quantities; and a data frame \code{theta}
with the intrinsic population growth rate \code{r}, carrying capacity
\code{k}, binomial likelihood \code{ell} (filtered to include only those with
a value of \code{1}, and the starting biomass depletion \code{J}. In the
biomass matrix, each row contains an iteration and each column contains a
year.
}
\description{
Apply the catch-MSY method from Martell and Froese (2013).
}
\examples{
# An example using cmsy() with a stock from the RAM Legacy database:
# The stock is "BGRDRSE" (Blue Grenadier Southeast Australia)

x <- cmsy(blue_gren$yr, ct = blue_gren$ct, reps = 2e4)
head(x$theta)
par(mfrow = c(2, 2))
plot(blue_gren$yr, blue_gren$ct, type = "o", xlab = "Year", ylab = "Catch (t)")
plot(blue_gren$yr,  apply(x$biomass, 2, median)[-1], type = "o",
  ylab = "Estimated biomass", xlab = "Year")
hist(x$bmsy)
plot(x$theta$r, x$theta$k)
bbmsy <- x$biomass[, -1] / x$bmsy
bbmsy_out <- summarize_bbmsy(bbmsy)
bbmsy_out$year <- blue_gren$yr
library("ggplot2")
ggplot(bbmsy_out, aes(year, bbmsy_q50)) + geom_line()  +
    geom_ribbon(aes(ymin = bbmsy_q25, ymax = bbmsy_q75), alpha = 0.2) +
    geom_ribbon(aes(ymin = bbmsy_q2.5, ymax = bbmsy_q97.5), alpha = 0.1) +
    geom_hline(yintercept = 1, lty = 2)
}
\references{
Martell, S., & Froese, R. (2013). A simple method for estimating MSY from
catch and resilience. Fish and Fisheries, 14(4), 504-514.
\url{http://doi.org/10.1111/j.1467-2979.2012.00485.x}
}

