% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cmsy.R
\name{cmsy}
\alias{cmsy}
\title{Catch-MSY method}
\usage{
cmsy(yr, ct, prior_log_mean, prior_log_sd, interyr_index = 2L,
  interbio = c(0, 1), bio_step = 0.05, start_r = resilience(NA),
  start_k = c(max(ct), 50 * max(ct)), startbio = if (ct[1]/max(ct) < 0.2)
  c(0.5, 0.9) else c(0.2, 0.6), sig_r = 0.05, reps = 10000,
  remove_ell0 = TRUE)
}
\arguments{
\item{yr}{Numeric vector of years}

\item{ct}{Numeric vector of catch}

\item{prior_log_mean}{Prior log mean}

\item{prior_log_sd}{Prior log sd}

\item{interyr_index}{Index of the interim year within time series for which
biomass estimate is available}

\item{interbio}{A numeric vector that gives the lower and upper biomass
limits in the interim year}

\item{bio_step}{Step size between lower and upper biomass limits in the iterim year}

\item{start_r}{A numeric vector of length 2 giving the lower and upper
bounds on the population growth rate parameter. This can either be
specified manually or by translating resiliency categories via the function
\code{\link{resilience}}}

\item{start_k}{Numeric vector of length 2 giving the lower and upper starting
bounds on stock biomass at carrying capacity}

\item{startbio}{Starting biomass}

\item{sig_r}{Standard deviation of process noise}

\item{reps}{Number of repititions to run the sampling}

\item{remove_ell0}{Should sampled rows with \code{ell == 0} be removed
internally?}
}
\value{
A data frame containing columns for intrinsic population growth rates
  (\code{r}), carrying capacity (\code{k}), log likelihood (\code{ell}), and
  biomass (\code{biomass}). Each row contains an iteration for a total length
  of \code{reps}.
}
\description{
Apply the catch-MSY method from Martell and Froese (2013).
}
\examples{
# An example using cmsy() with a stock from the RAM Legacy database:
# The stock is "BGRDRSE" (Blue Grenadier Southeast Australia)

x <- cmsy(blue_gren$yr, ct = blue_gren$ct, prior_log_mean = 0.035,
  prior_log_sd = 0.68)
head(x$shaefer)
par(mfrow = c(2, 2))
plot(blue_gren$yr, blue_gren$ct, type = "o", xlab = "Year", ylab = "Catch (t)")
plot(blue_gren$yr,  apply(x$biomass, 1, median)[-1], type = "o",
  ylab = "Estimated biomass", xlab = "Year")
hist(x$shaefer$bmsy)
plot(x$shaefer$r, x$shaefer$k)
}
\references{
Martell, S., & Froese, R. (2013). A simple method for estimating MSY from
catch and resilience. Fish and Fisheries, 14(4), 504-514.
\url{http://doi.org/10.1111/j.1467-2979.2012.00485.x}
}

