% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comsir.R
\name{comsir}
\alias{comsir}
\title{COM-SIR method}
\description{
COM-SIR method
}
\examples{
out <- comsir(Catch = rlnorm(10), yrs = 1:10, K = 100, r = 0.1, x = 1, a = 1,
  start_r = c(0.05, 0.2))
head(out)
par(mfrow = c(1, 2))
hist(out$BoverBmsy)
with(out$posterior, plot(r, K))
}

