% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comsir.R
\name{comsir}
\alias{comsir}
\title{COM-SIR method}
\arguments{
\item{yr}{A time series of years associated with the catch}

\item{ct}{A time series of catch}

\item{start_r}{A numeric vector of length 2 giving the lower and upper
bounds on the population growth rate parameter. This can either be
specified manually or by translating resiliency categories via the function
\code{\link{resilience}}.}

\item{k_bounds}{Minimum and maximum possible stock size at carrying capacity}

\item{x_bounds}{Minimum and maximum possible x (effort dynamics parameter) values}

\item{a_bounds}{Minimum and maximum possible a (effort dynamics parameter) values}

\item{norm_k}{Logical: should \code{k} have a normal prior (\code{TRUE}) or a
uniform prior between \code{mink} and \code{maxk} (\code{FALSE})?}

\item{logk}{If \code{logk = TRUE} and \code{norm_k = FALSE} the prior on
\code{k} will be an exponentiated form a uniform distribution on a log
scale between \code{log(mink)} and \code{log(maxk)}.}

\item{nsim}{Number of iterations to run before sampling}

\item{logistic_model}{Logical: if \code{TRUE} then the effort dynamics model
will be a logistic model. If \code{FALSE} then the effort dynamics model will
be linear.}

\item{n_posterior}{Number of posterior samples to draw}

\item{obs}{Logical: if \code{FALSE} then a measurement-error catch model is}

\item{effort_bounds}{Lower and upper limits on rate of decrease and increase
in effort from one time step to the next.}

\item{dampen}{Should effort dynamics parameters be excluded that may give
unstable dynamics?}

\item{cv_bounds}{Min and max limits on residual error CV around catch}
}
\description{
Catch-only model with sample importance resampling based on the method
described in Vasconcellos and Cochrane (2005).
}
\examples{
# TODO K and r values?
x <- comsir(ct = blue_gren$ct, yr = blue_gren$yr, nsim = 1e5,
  n_posterior = 2e3)
par(mfrow = c(1, 2))
hist(x$quantities$bbmsy)
with(x$posterior, plot(r, k))
}
\references{
Vasconcellos, M., and K. Cochrane. 2005. Overview of World Status of
Data-Limited Fisheries: Inferences from Landings Statistics. Pages 1-20 in G.
H. Kruse, V. F. Gallucci, D. E. Hay, R. I. Perry, R. M. Peterman, T. C.
Shirley, P. D. Spencer, B. Wilson, and D. Woodby, editors. Fisheries
Assessment and Management in Data-Limited Situations. Alaska Sea Grant,
University of Alaska Fairbanks.
}

