% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prm.R
\name{prm}
\alias{fit_prm}
\alias{format_prm}
\alias{predict_prm}
\alias{prm}
\title{Panel regression model (Costello et al.-style model)}
\usage{
format_prm(year, catch, bbmsy, species_cat)

fit_prm(dat)

predict_prm(newdata, model = datalimited::ram_prm_model, ci = FALSE,
  level = 0.95)
}
\arguments{
\item{year}{A numeric vector of years}

\item{catch}{A numeric vector of catches}

\item{bbmsy}{A numeric vector of B/B_MSY}

\item{species_cat}{A single character value of a
species category. Can be from any set of species categories as long as the
same set are used in model fitting and extrapolation.}

\item{dat}{A data frame created by \code{\link{format_prm}}}

\item{newdata}{A data frame to predict on that has been formatted with
\code{\link{format_prm}}.}

\item{model}{A linear model to predict from built from \code{\link{fit_prm}}.
Defaults to a cached version of a model fit to the RAM database,
\code{\link{ram_prm_model}}.}

\item{ci}{Should confidence intervals on B/B_MSY be returned?}

\item{level}{Confidence interval level}
}
\value{
\code{format_prm}: A data frame formatted for use with \code{fit_prm} or
  \code{predict_prm}

\code{fit_prm}: A linear model for use with \code{predict_prm}

\code{predict_prm}: If \code{ci = FALSE}, a vector of predictions of
  B/B_MSY. If \code{ci = TRUE}, a data frame with columns for B/B_MSY
  \code{fit} and lower and upper confidence intervals \code{lower},
  \code{upper}. Note that the model is fitted to log(B/B_MSY) and the output
  from \code{predict_prm} is exponentiated so the prediction represents an
  estimate of median B/B_MSY.
}
\description{
Format for, fit, and predict from a panel regression model. The
\pkg{datalimited} package includes cached data that can be used with
\code{fit_prm} and a cached model that can be used with \code{predict_prm}.
These functions can also be used with new data.

\code{format_prm}: Format a series of predictors for a panel
  regression

\code{fit_prm}: Fit a panel regression

\code{predict_prm}: Predict from a panel regression model
}
\examples{
# format_prm()
d <- dplyr::inner_join(ram_ts, spp_categories, by = "scientificname")
ram_prm_dat <- plyr::ddply(d, "stockid", function(x) {
  format_prm(year = x$year, catch = x$catch, bbmsy = x$bbmsy_ram,
    species_cat = x$spp_category[1L])
})

# fit_prm()
m <- fit_prm(ram_prm_dat)

# predict_prm()
d <- subset(ram_prm_dat, stockid == "BGRDRSE")
x <- predict_prm(d, model = m)
plot(x)

library("ggplot2")
x <- predict_prm(d, ci = TRUE)
x$year <- rev(d$years_back)
ggplot(x, aes(year, fit)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), fill = "#00000040") +
  geom_line() +
  ylab(expression(B/B[MSY]))
}
\references{
Costello, C., D. Ovando, R. Hilborn, S. D. Gaines, O. Deschenes,
and S. E. Lester. 2012. Status and Solutions for the World's Unassessed
Fisheries. Science 338:517-520.
}

