% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ram_data.R
\docType{data}
\name{spp_categories}
\alias{spp_categories}
\title{Species names and categories for use with the RAM database}
\description{
RAM database cached from version 2.5.
}
\references{
Ricard, D., Minto, C., Jensen, O. P., & Baum, J. K. (2012). Examining the
knowledge base and status of commercially exploited marine species with the
RAM Legacy Stock Assessment Database. Fish and Fisheries, 13(4), 3 80-398.
\url{http://doi.org/10.1111/j.1467-2979.2011.00435.x}.
\url{http://ramlegacy.org}
}
\keyword{data}

