% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sscom.R
\name{sscom}
\alias{sscom}
\title{SSCOM model}
\usage{
sscom(Year, Ct, start_r = resilience("unknown"), NburninPrelim = 1000,
  NiterPrelim = 2000, NthinPrelim = 1, NchainsPrelim = 100,
  NburninJags = 1e+06, NiterJags = 2e+06, NthinJags = 1000, Nchains = 3,
  return_jags = FALSE)
}
\arguments{
\item{Year}{Years
NburninPrelim = 1e3,}

\item{Ct}{Catch time series}

\item{start_r}{Lower and upper intrinsic population growth rates}

\item{NiterPrelim}{Number of preliminary iterations}

\item{NthinPrelim}{Thinning of preliminary iterations}

\item{NchainsPrelim}{Number of chains for preliminary iterations}

\item{NburninJags}{Number of warmup iterations for main sampling phase}

\item{NiterJags}{Number of sampling iterations for main sampling phase}

\item{NthinJags}{Thining for main sampling phase}

\item{return_jags}{Logical: should the JAGS output object be returned at the
end as well. If \code{TRUE} then a list with elements \code{jags} and
\code{output} will be returned with the JAGS and SSCOM output, respectively.}
}
\description{
SSCOM model
}

